move python2/3 code into utils.py
