+++ aggregate/aggregate_numpy.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from .utils import (check_boolean, _no_separate_nan_version, get_func,
                     aliasing, fill_untouched, minimum_dtype, input_validation,
-                    check_dtype, minimum_dtype_scalar, _doc_str)
+                    check_dtype, minimum_dtype_scalar, _doc_str, isstr)
 
 
 def _sort(group_idx, a, size, fill_value, dtype=None, reversed_=False):
@@ -212,16 +212,7 @@ _impl_dict = dict(min=_min, max=_max, sum=_sum, prod=_prod, last=_last,
 _impl_dict.update(('nan' + k, v) for k, v in list(_impl_dict.items())
                   if k not in _no_separate_nan_version)
 
-
-try:
-    basestring  # attempt to evaluate basestring
-    def isstr(s):
-        return isinstance(s, basestring)
-except NameError:
-    # probably Python 3.x
-    def isstr(s):
-        return isinstance(s, str)
-        
+       
 def _aggregate_base(group_idx, a, func='sum', size=None, fill_value=0,
                     order='C', dtype=None, _impl_dict=_impl_dict,
                     _nansqueeze=False, **kwargs):
