+++ aggregate/aggregate_weave.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.weave import inline
 
 from .utils import (_no_separate_nan_version, get_func, check_dtype,
-                    aliasing, check_fill_value, input_validation, _doc_str)
+                    aliasing, check_fill_value, input_validation, _doc_str, isstr)
 from .aggregate_numpy import aggregate as aggregate_np
 
 
@@ -213,14 +213,6 @@ def step_indices(group_idx):
     inline(c_step_indices, ['group_idx', 'indices'], define_macros=c_macros)
     return indices
 
-try:
-    basestring  # attempt to evaluate basestring
-    def isstr(s):
-        return isinstance(s, basestring)
-except NameError:
-    # probably Python 3.x
-    def isstr(s):
-        return isinstance(s, str)
         
 def aggregate(group_idx, a, func='sum', size=None, fill_value=0, order='C', dtype=None, **kwargs):
     func = get_func(func, aliasing, optimized_funcs)
